requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-community==0.2.6
langchain-chroma==0.1.2
chromadb==0.5.3
sentence-transformers==3.0.1
python-dotenv==1.0.1
httpx==0.27.0
python-docx==1.1.2
